chore: app files reorganized to better rep app
