+++ src/App.js
@@ -1,27 +1,23 @@
-import React from 'react';
-import Navbar from "./components/navbar";
+import React from "react";
+import Navbar from "./components/parts/navbar/navbar";
 import { BrowserRouter, Route } from "react-router-dom";
-import Home from './pages/home';
-import Portfolio from './pages/portfolio';
-import About from './pages/about';
-import Contact from './pages/contact'
+import Home from "./pages/home/home";
+import Portfolio from "./pages/portfolio/portfolio";
+import About from "./components/details/about/about";
+import Contact from "./pages/contact/contact";
 
 class App extends React.Component {
-  state = {
-  }
+  state = {};
 
   render() {
     return (
       <BrowserRouter>
-     
         <Navbar />
-        <Route exact path='/' component={Home} />
-        <Route path='/portfolio' component={Portfolio}/>
-        <Route exact path='/' component={About} />
-        <Route path='/contact' component={Contact} />
-        
+        <Route exact path="/" component={Home} />
+        <Route path="/portfolio" component={Portfolio} />
+        <Route exact path="/" component={About} />
+        <Route path="/contact" component={Contact} />
       </BrowserRouter>
-
     );
   }
 }
