+++ src/components/details/about/about.css
