+++ src/components/parts/technologies/techsComp.css
