+++ src/components/parts/technologies/techsComp.js
@@ -1,0 +1,11 @@
+import React from "react";
+import "./techsComp.css";
+
+const TechsComp = (props) => {
+  return (
+    <li key={props.technology} id={props.id}>
+      <img className="techlogo" alt={props.technology} src={props.image} />
+    </li>
+  );
+};
+export default TechsComp;
