--- src/pages/home.js
@@ -1,20 +1,0 @@
-import React from "react";
-import './home.css';
-
-const Home = () => {
-    return (
-        <div id="landing-section">
-            <figure id='thumbnail'>
-                <img id='woodsImage' alt='woods' src='../../images/woods2_copy.png'></img>
-            </figure>
-            <div className="text-container">
-                <p className="static-text">Hi, I'm Camerone</p>
-            </div>
-            <div id='overlay'></div>
-           
-        </div>
-
-    )
-}
-
-export default Home;
